Refactoring: Octohax, Sisterhax and Unlocks logic extracted. - Octohax and Sisterhax tested, still work.
