+++ SplatAIO/Logic/Hacks/Sisterhax/SisterhaxAddress.cs
@@ -1,0 +1,12 @@
+﻿namespace SplatAIO.Logic.Memory.Addresses
+{
+    public enum SisterhaxAddress : uint
+    {
+        Aori = 0x105EB5EC, // Callie
+        AoriTwo = Aori + 0x4,
+        AoriModel = Aori + 0x8,
+        Hotaru = Aori + 0xC, //0x105EB5F8 // Marie
+        HotaruTwo = Hotaru + 0x4,
+        HotaruModel = Hotaru + 0x8
+    }
+}
