+++ SplatAIO/Logic/Hacks/Sisterhax/SisterhaxValue.cs
@@ -1,0 +1,10 @@
+﻿namespace SplatAIO.Logic.Hacks.Sisterhax
+{
+    public enum SisterhaxValue : uint
+    {
+        AoriModel = 0x41000000, // Aori's Model 
+        HotaruModel = 0x42000000, // Hotaru's Model
+        ValueOne = 0x4E70635F, // I don't know what are this values
+        ValueTwo = 0x49646F6C // They are always the same in Sisterhax
+    }
+}
