+++ SplatAIO/Logic/Hacks/Unlock/GearAddress.cs
@@ -1,6 +1,6 @@
-﻿namespace SplatAIO.Logic.Memory.Addresses
+﻿namespace SplatAIO.Logic.Hacks.Unlock
 {
-    public enum Gear : uint
+    public enum GearAddress : uint
     {
         EquippedClothes = 0x12CD1D80,
         EquippedShoes = EquippedClothes + 0x4, // 0x12CD1D84
