+++ SplatAIO/Logic/Hacks/Unlock/GearUnlocker.cs
@@ -1,0 +1,333 @@
+﻿using SplatAIO.Logic.Gecko;
+using System.Collections.Generic;
+using System.Linq;
+
+namespace SplatAIO.Logic.Hacks.Unlock
+{
+    public class GearUnlocker
+    {
+        private readonly Dictionary<uint, uint[]> _clothes = new Dictionary<uint, uint[]>
+        {
+            {0x00000001, new uint[] {0x00000000, 0x00000000, 0x00000000}},
+            {0x000003e8, new uint[] {0x00000000, 0x00000000, 0x00000000}},
+            {0x000003e9, new uint[] {0x00000004, 0x00000004, 0x00000004}},
+            {0x000003eb, new uint[] {0x00000004, 0x00000004, 0x00000004}},
+            {0x000003ec, new uint[] {0x00000005, 0x00000005, 0x00000005}},
+            {0x000003ed, new uint[] {0x00000005, 0x00000005, 0x00000005}},
+            {0x000003ee, new uint[] {0x00000000, 0x00000000, 0x00000000}},
+            {0x000003ef, new uint[] {0x00000006, 0x00000006, 0x00000006}},
+            {0x000003f0, new uint[] {0x00000006, 0x00000006, 0x00000006}},
+            {0x000003f1, new uint[] {0x00000009, 0x00000009, 0x00000009}},
+            {0x000003f2, new uint[] {0x00000003, 0x00000003, 0x00000003}},
+            {0x000003f3, new uint[] {0x00000009, 0x00000009, 0x00000009}},
+            {0x000003f4, new uint[] {0x00000009, 0x00000009, 0x00000009}},
+            {0x000003f5, new uint[] {0x00000007, 0x00000007, 0x00000007}},
+            {0x000003f6, new uint[] {0x00000007, 0x00000007, 0x00000007}},
+            {0x000003f7, new uint[] {0x00000001, 0x00000001, 0x00000001}},
+            {0x000003f8, new uint[] {0x00000001, 0x00000001, 0x00000001}},
+            {0x000003f9, new uint[] {0x00000009, 0x00000009, 0x00000009}},
+            {0x000003fa, new uint[] {0x00000002, 0x00000002, 0x00000002}},
+            {0x000003fb, new uint[] {0x00000002, 0x00000002, 0x00000002}},
+            {0x000003fc, new uint[] {0x00000003, 0x00000003, 0x00000003}},
+            {0x000003fd, new uint[] {0x00000003, 0x00000003, 0x00000003}},
+            {0x000003fe, new uint[] {0x00000000, 0x00000000, 0x00000000}},
+            {0x000003ff, new uint[] {0x00000000, 0x00000000, 0x00000000}},
+            {0x00000402, new uint[] {0x00000003, 0x00000003, 0x00000003}},
+            {0x00000403, new uint[] {0x00000009, 0x00000009, 0x00000009}},
+            {0x00000405, new uint[] {0x00000003, 0x00000003, 0x00000003}},
+            {0x000007d0, new uint[] {0x00000002, 0x00000002, 0x00000002}},
+            {0x000007d1, new uint[] {0x0000000A, 0x0000000A, 0x0000000A}},
+            {0x000007d2, new uint[] {0x00000007, 0x00000007, 0x00000007}},
+            {0x000007d3, new uint[] {0x00000002, 0x00000002, 0x00000002}},
+            {0x000007d4, new uint[] {0x0000000A, 0x0000000A, 0x0000000A}},
+            {0x000007d5, new uint[] {0x00000002, 0x00000002, 0x00000002}},
+            {0x000007d6, new uint[] {0x00000005, 0x00000005, 0x00000005}},
+            {0x000007d7, new uint[] {0x00000005, 0x00000005, 0x00000005}},
+            {0x000007d8, new uint[] {0x00000000, 0x00000000, 0x00000000}},
+            {0x000007d9, new uint[] {0x00000001, 0x00000001, 0x00000001}},
+            {0x000007da, new uint[] {0x00000000, 0x00000000, 0x00000000}},
+            {0x000007db, new uint[] {0x0000000B, 0x0000000B, 0x0000000B}},
+            {0x000007dc, new uint[] {0x00000001, 0x00000001, 0x00000001}},
+            {0x00000bb8, new uint[] {0x00000000, 0x00000000, 0x00000000}},
+            {0x00000bb9, new uint[] {0x00000000, 0x00000000, 0x00000000}},
+            {0x00000bba, new uint[] {0x00000000, 0x00000000, 0x00000000}},
+            {0x00000bbb, new uint[] {0x00000000, 0x00000000, 0x00000000}},
+            {0x00000bbc, new uint[] {0x0000000B, 0x0000000B, 0x0000000B}},
+            {0x00000bbd, new uint[] {0x00000000, 0x00000000, 0x00000000}},
+            {0x00000bbe, new uint[] {0x00000007, 0x00000007, 0x00000007}},
+            {0x00000bbf, new uint[] {0x00000004, 0x00000004, 0x00000004}},
+            {0x00000bc0, new uint[] {0x00000007, 0x00000007, 0x00000007}},
+            {0x00000bc1, new uint[] {0x00000008, 0x00000008, 0x00000008}},
+            {0x00000fa0, new uint[] {0x00000002, 0x00000002, 0x00000002}},
+            {0x00000fa1, new uint[] {0x00000007, 0x00000007, 0x00000007}},
+            {0x00000fa2, new uint[] {0x00000007, 0x00000007, 0x00000007}},
+            {0x00000fa3, new uint[] {0x00000002, 0x00000002, 0x00000002}},
+            {0x00000fa4, new uint[] {0x0000000A, 0x0000000A, 0x0000000A}},
+            {0x00000fa5, new uint[] {0x0000000B, 0x0000000B, 0x0000000B}},
+            {0x00000fa6, new uint[] {0x00000004, 0x00000004, 0x00000004}},
+            {0x00000fa7, new uint[] {0x00000007, 0x00000007, 0x00000007}},
+            {0x00000fa8, new uint[] {0x00000007, 0x00000007, 0x00000007}},
+            {0x00001388, new uint[] {0x00000001, 0x00000001, 0x00000001}},
+            {0x0000138a, new uint[] {0x00000001, 0x00000001, 0x00000001}},
+            {0x0000138b, new uint[] {0x0000000A, 0x0000000A, 0x0000000A}},
+            {0x0000138c, new uint[] {0x0000000B, 0x0000000B, 0x0000000B}},
+            {0x0000138d, new uint[] {0x00000002, 0x00000002, 0x00000002}},
+            {0x0000138e, new uint[] {0x00000005, 0x00000005, 0x00000005}},
+            {0x0000138f, new uint[] {0x00000005, 0x00000005, 0x00000005}},
+            {0x00001390, new uint[] {0x0000000B, 0x0000000B, 0x0000000B}},
+            {0x00001391, new uint[] {0x00000002, 0x00000002, 0x00000002}},
+            {0x00001392, new uint[] {0x00000008, 0x00000008, 0x00000008}},
+            {0x00001393, new uint[] {0x00000008, 0x00000008, 0x00000008}},
+            {0x00001394, new uint[] {0x00000008, 0x00000008, 0x00000008}},
+            {0x00001395, new uint[] {0x00000008, 0x00000008, 0x00000008}},
+            {0x00001396, new uint[] {0x0000000A, 0x0000000A, 0x0000000A}},
+            {0x00001397, new uint[] {0x0000000A, 0x0000000A, 0x0000000A}},
+            {0x00001398, new uint[] {0x00000006, 0x00000006, 0x00000006}},
+            {0x00001770, new uint[] {0x0000000B, 0x0000000B, 0x0000000B}},
+            {0x00001771, new uint[] {0x0000000B, 0x0000000B, 0x0000000B}},
+            {0x00001b58, new uint[] {0x00000002, 0x00000002, 0x00000002}},
+            {0x00001b59, new uint[] {0x00000000, 0x00000000, 0x00000000}},
+            {0x00001b5a, new uint[] {0x00000000, 0x00000000, 0x00000000}},
+            {0x00001b5b, new uint[] {0x00000003, 0x00000003, 0x00000003}},
+            {0x00001b5c, new uint[] {0x00000002, 0x00000002, 0x00000002}},
+            {0x00001b5d, new uint[] {0x0000000A, 0x0000000A, 0x0000000A}},
+            {0x00001b5e, new uint[] {0x00000000, 0x00000000, 0x00000000}},
+            {0x00001f40, new uint[] {0x00000005, 0x00000005, 0x00000005}},
+            {0x00001f41, new uint[] {0x00000006, 0x00000006, 0x00000006}},
+            {0x00001f42, new uint[] {0x0000000A, 0x0000000A, 0x0000000A}},
+            {0x00001f43, new uint[] {0x00000002, 0x00000002, 0x00000002}},
+            {0x00001f44, new uint[] {0x0000000B, 0x0000000B, 0x0000000B}},
+            {0x00001f45, new uint[] {0x00000008, 0x00000008, 0x00000008}},
+            {0x00001f46, new uint[] {0x0000000A, 0x0000000A, 0x0000000A}},
+            {0x00001f47, new uint[] {0x00000002, 0x00000002, 0x00000002}},
+            {0x00001f48, new uint[] {0x00000003, 0x00000003, 0x00000003}},
+            {0x00001f49, new uint[] {0x0000000A, 0x0000000A, 0x0000000A}},
+            {0x00001f4a, new uint[] {0x00000005, 0x00000005, 0x00000005}},
+            {0x00001f4b, new uint[] {0x00000005, 0x00000005, 0x00000005}},
+            {0x00001f4c, new uint[] {0x0000000A, 0x0000000A, 0x0000000A}},
+            {0x00001f4d, new uint[] {0x00000002, 0x00000002, 0x00000002}},
+            {0x00001f4e, new uint[] {0x00000002, 0x00000002, 0x00000002}},
+            {0x00001f4f, new uint[] {0x00000002, 0x00000002, 0x00000002}},
+            {0x00002328, new uint[] {0x00000001, 0x00000001, 0x00000001}},
+            {0x00002329, new uint[] {0x00000001, 0x00000001, 0x00000001}},
+            {0x0000232a, new uint[] {0x00000003, 0x00000003, 0x00000003}},
+            {0x0000232b, new uint[] {0x00000003, 0x00000003, 0x00000003}},
+            {0x0000232c, new uint[] {0x00000006, 0x00000006, 0x00000006}},
+            {0x0000232d, new uint[] {0x00000006, 0x00000006, 0x00000006}},
+            {0x00002710, new uint[] {0x00000003, 0x00000003, 0x00000003}},
+            {0x00002711, new uint[] {0x00000003, 0x00000003, 0x00000003}},
+            {0x00002712, new uint[] {0x0000000A, 0x0000000A, 0x0000000A}},
+            {0x00000400, new uint[] {0x00000006, 0x00000005, 0x00000005}},
+            {0x00000401, new uint[] {0x0000000A, 0x0000000C, 0x00000001}},
+            {0x00001772, new uint[] {0x00000000, 0x00000005, 0x00000001}},
+            {0x00001f50, new uint[] {0x00000004, 0x0000000C, 0x0000000B}},
+            {0x000061a8, new uint[] {0x0000000B, 0x0000000A, 0x0000000C}},
+            {0x000061a9, new uint[] {0x00000007, 0x00000008, 0x00000008}},
+            {0x000061aa, new uint[] {0x00000009, 0x00000001, 0x0000000C}},
+            {0x00000404, new uint[] {0x0000000C, 0x0000000C, 0x00000009}},
+            {0x00006978, new uint[] {0x0000000B, 0x0000000C, 0x0000000A}},
+            {0x0000697c, new uint[] {0x00000008, 0x00000003, 0x00000006}},
+            {0x00006d60, new uint[] {0x0000000C, 0x00000003, 0x0000000C}},
+            {0x00002713, new uint[] {0x0000000A, 0x0000000A, 0x0000000A}}
+        };
+
+        private readonly Dictionary<uint, uint[]> _hats = new Dictionary<uint, uint[]>
+        {
+            {0x00000001, new uint[] {0x00000000, 0x00000000, 0x00000000}},
+            {0x000003e8, new uint[] {0x00000009, 0x00000009, 0x00000009}},
+            {0x000003e9, new uint[] {0x00000001, 0x00000001, 0x00000001}},
+            {0x000003ea, new uint[] {0x00000007, 0x00000007, 0x00000007}},
+            {0x000003eb, new uint[] {0x00000009, 0x00000009, 0x00000009}},
+            {0x000003ec, new uint[] {0x00000009, 0x00000009, 0x00000009}},
+            {0x000003ed, new uint[] {0x00000009, 0x00000009, 0x00000009}},
+            {0x000003ee, new uint[] {0x00000003, 0x00000003, 0x00000003}},
+            {0x000003ef, new uint[] {0x0000000A, 0x0000000A, 0x0000000A}},
+            {0x000003f0, new uint[] {0x0000000A, 0x0000000A, 0x0000000A}},
+            {0x000003f1, new uint[] {0x0000000A, 0x0000000A, 0x0000000A}},
+            {0x000003f2, new uint[] {0x00000006, 0x00000006, 0x00000006}},
+            {0x000003f3, new uint[] {0x00000003, 0x00000003, 0x00000003}},
+            {0x000003f4, new uint[] {0x0000000B, 0x0000000B, 0x0000000B}},
+            {0x000003f6, new uint[] {0x00000004, 0x00000004, 0x00000004}},
+            {0x000007d0, new uint[] {0x00000002, 0x00000002, 0x00000002}},
+            {0x000007d1, new uint[] {0x00000001, 0x00000001, 0x00000001}},
+            {0x000007d2, new uint[] {0x00000002, 0x00000002, 0x00000002}},
+            {0x000007d3, new uint[] {0x00000009, 0x00000009, 0x00000009}},
+            {0x000007d4, new uint[] {0x00000008, 0x00000008, 0x00000008}},
+            {0x000007d5, new uint[] {0x00000009, 0x00000009, 0x00000009}},
+            {0x00000bb8, new uint[] {0x00000002, 0x00000002, 0x00000002}},
+            {0x00000bb9, new uint[] {0x00000008, 0x00000008, 0x00000008}},
+            {0x00000bba, new uint[] {0x00000008, 0x00000008, 0x00000008}},
+            {0x00000bbb, new uint[] {0x0000000A, 0x0000000A, 0x0000000A}},
+            {0x00000bbc, new uint[] {0x0000000A, 0x0000000A, 0x0000000A}},
+            {0x00000bbd, new uint[] {0x00000008, 0x00000008, 0x00000008}},
+            {0x00000bbe, new uint[] {0x0000000A, 0x0000000A, 0x0000000A}},
+            {0x00000bbf, new uint[] {0x00000004, 0x00000004, 0x00000004}},
+            {0x00000bc0, new uint[] {0x00000005, 0x00000005, 0x00000005}},
+            {0x00000bc1, new uint[] {0x00000006, 0x00000006, 0x00000006}},
+            {0x00000bc2, new uint[] {0x00000003, 0x00000003, 0x00000003}},
+            {0x00000fa0, new uint[] {0x00000008, 0x00000008, 0x00000008}},
+            {0x00000fa1, new uint[] {0x00000003, 0x00000003, 0x00000003}},
+            {0x00000fa2, new uint[] {0x00000001, 0x00000001, 0x00000001}},
+            {0x00000fa3, new uint[] {0x00000003, 0x00000003, 0x00000003}},
+            {0x00000fa4, new uint[] {0x00000001, 0x00000001, 0x00000001}},
+            {0x00000fa5, new uint[] {0x00000009, 0x00000009, 0x00000009}},
+            {0x00000fa6, new uint[] {0x00000009, 0x00000009, 0x00000009}},
+            {0x00000fa7, new uint[] {0x00000008, 0x00000008, 0x00000008}},
+            {0x00001388, new uint[] {0x00000008, 0x00000008, 0x00000008}},
+            {0x00001389, new uint[] {0x00000008, 0x00000008, 0x00000008}},
+            {0x0000138a, new uint[] {0x00000008, 0x00000008, 0x00000008}},
+            {0x00001770, new uint[] {0x0000000B, 0x0000000B, 0x0000000B}},
+            {0x00001771, new uint[] {0x00000003, 0x00000003, 0x00000003}},
+            {0x00001772, new uint[] {0x00000004, 0x00000004, 0x00000004}},
+            {0x00001b58, new uint[] {0x00000009, 0x00000009, 0x00000009}},
+            {0x00001b5a, new uint[] {0x00000008, 0x00000008, 0x00000008}},
+            {0x00001b5b, new uint[] {0x00000008, 0x00000008, 0x00000008}},
+            {0x00001b5c, new uint[] {0x00000009, 0x00000009, 0x00000009}},
+            {0x00001b5d, new uint[] {0x00000009, 0x00000009, 0x00000009}},
+            {0x00001f40, new uint[] {0x00000008, 0x00000008, 0x00000008}},
+            {0x00001f41, new uint[] {0x00000008, 0x00000008, 0x00000008}},
+            {0x00001f42, new uint[] {0x00000006, 0x00000006, 0x00000006}},
+            {0x00001f43, new uint[] {0x00000008, 0x00000008, 0x00000008}},
+            {0x00002329, new uint[] {0x0000000B, 0x0000000B, 0x0000000B}},
+            {0x0000232a, new uint[] {0x0000000B, 0x0000000B, 0x0000000B}},
+            {0x0000232b, new uint[] {0x00000004, 0x00000004, 0x00000004}},
+            {0x0000232c, new uint[] {0x0000000A, 0x0000000A, 0x0000000A}},
+            {0x0000232d, new uint[] {0x00000004, 0x00000004, 0x00000004}},
+            {0x000003f5, new uint[] {0x00000005, 0x00000006, 0x0000000C}},
+            {0x0000232e, new uint[] {0x0000000A, 0x0000000C, 0x00000003}},
+            {0x000061a8, new uint[] {0x00000006, 0x00000004, 0x00000005}},
+            {0x000061a9, new uint[] {0x00000000, 0x00000001, 0x00000005}},
+            {0x000061aa, new uint[] {0x00000001, 0x00000001, 0x0000000B}},
+            {0x00006978, new uint[] {0x00000002, 0x00000005, 0x00000006}},
+            {0x0000697c, new uint[] {0x00000007, 0x00000007, 0x00000008}},
+            {0x00006d60, new uint[] {0x0000000C, 0x0000000C, 0x00000003}},
+            {0x000003f7, new uint[] {0x00000000, 0x00000001, 0x00000005}},
+            {0x000003f8, new uint[] {0x0000000A, 0x0000000A, 0x0000000A}}
+        };
+
+        private readonly Dictionary<uint, uint[]> _shoes = new Dictionary<uint, uint[]>
+        {
+            {0x00000001, new uint[] {0x00000006, 0x00000006, 0x00000006}},
+            {0x000003e8, new uint[] {0x0000000A, 0x0000000A, 0x0000000A}},
+            {0x000003e9, new uint[] {0x00000006, 0x00000006, 0x00000006}},
+            {0x000003ea, new uint[] {0x0000000A, 0x0000000A, 0x0000000A}},
+            {0x000003eb, new uint[] {0x0000000B, 0x0000000B, 0x0000000B}},
+            {0x000003ec, new uint[] {0x0000000A, 0x0000000A, 0x0000000A}},
+            {0x000003ed, new uint[] {0x0000000B, 0x0000000B, 0x0000000B}},
+            {0x000003ee, new uint[] {0x00000006, 0x00000006, 0x00000006}},
+            {0x000003ef, new uint[] {0x0000000B, 0x0000000B, 0x0000000B}},
+            {0x000003f0, new uint[] {0x00000002, 0x00000002, 0x00000002}},
+            {0x000003f1, new uint[] {0x00000002, 0x00000002, 0x00000002}},
+            {0x000003f2, new uint[] {0x00000007, 0x00000007, 0x00000007}},
+            {0x000003f3, new uint[] {0x00000007, 0x00000007, 0x00000007}},
+            {0x000007d0, new uint[] {0x0000000B, 0x0000000B, 0x0000000B}},
+            {0x000007d1, new uint[] {0x0000000B, 0x0000000B, 0x0000000B}},
+            {0x000007d2, new uint[] {0x00000006, 0x00000006, 0x00000006}},
+            {0x000007d3, new uint[] {0x0000000B, 0x0000000B, 0x0000000B}},
+            {0x000007d4, new uint[] {0x00000006, 0x00000006, 0x00000006}},
+            {0x000007d5, new uint[] {0x00000006, 0x00000006, 0x00000006}},
+            {0x000007d6, new uint[] {0x0000000B, 0x0000000B, 0x0000000B}},
+            {0x000007d8, new uint[] {0x00000002, 0x00000002, 0x00000002}},
+            {0x000007d9, new uint[] {0x00000002, 0x00000002, 0x00000002}},
+            {0x00000bb8, new uint[] {0x00000004, 0x00000004, 0x00000004}},
+            {0x00000bb9, new uint[] {0x00000007, 0x00000007, 0x00000007}},
+            {0x00000bba, new uint[] {0x00000004, 0x00000004, 0x00000004}},
+            {0x00000bbb, new uint[] {0x00000007, 0x00000007, 0x00000007}},
+            {0x00000bbc, new uint[] {0x00000004, 0x00000004, 0x00000004}},
+            {0x00000bbd, new uint[] {0x00000004, 0x00000004, 0x00000004}},
+            {0x00000bbe, new uint[] {0x00000004, 0x00000004, 0x00000004}},
+            {0x00000bbf, new uint[] {0x00000004, 0x00000004, 0x00000004}},
+            {0x00000bc0, new uint[] {0x00000007, 0x00000007, 0x00000007}},
+            {0x00000bc1, new uint[] {0x00000004, 0x00000004, 0x00000004}},
+            {0x00000fa0, new uint[] {0x00000006, 0x00000006, 0x00000006}},
+            {0x00000fa1, new uint[] {0x00000006, 0x00000006, 0x00000006}},
+            {0x00000fa2, new uint[] {0x00000006, 0x00000006, 0x00000006}},
+            {0x00000fa3, new uint[] {0x00000006, 0x00000006, 0x00000006}},
+            {0x00001388, new uint[] {0x00000001, 0x00000001, 0x00000001}},
+            {0x00001389, new uint[] {0x00000001, 0x00000001, 0x00000001}},
+            {0x0000138a, new uint[] {0x00000001, 0x00000001, 0x00000001}},
+            {0x00001770, new uint[] {0x00000005, 0x00000005, 0x00000005}},
+            {0x00001771, new uint[] {0x00000005, 0x00000005, 0x00000005}},
+            {0x00001772, new uint[] {0x00000005, 0x00000005, 0x00000005}},
+            {0x00001773, new uint[] {0x00000005, 0x00000005, 0x00000005}},
+            {0x00001774, new uint[] {0x00000001, 0x00000001, 0x00000001}},
+            {0x00001775, new uint[] {0x00000001, 0x00000001, 0x00000001}},
+            {0x00001776, new uint[] {0x00000005, 0x00000005, 0x00000005}},
+            {0x00001777, new uint[] {0x00000005, 0x00000005, 0x00000005}},
+            {0x00001778, new uint[] {0x00000005, 0x00000005, 0x00000005}},
+            {0x00001779, new uint[] {0x00000001, 0x00000001, 0x00000001}},
+            {0x0000177a, new uint[] {0x00000004, 0x00000004, 0x00000004}},
+            {0x0000177b, new uint[] {0x00000004, 0x00000004, 0x00000004}},
+            {0x00001b58, new uint[] {0x00000006, 0x00000006, 0x00000006}},
+            {0x00001b59, new uint[] {0x00000006, 0x00000006, 0x00000006}},
+            {0x00001b5a, new uint[] {0x00000006, 0x00000006, 0x00000006}},
+            {0x00001f40, new uint[] {0x00000005, 0x00000005, 0x00000005}},
+            {0x00001f41, new uint[] {0x00000005, 0x00000005, 0x00000005}},
+            {0x00001f42, new uint[] {0x00000005, 0x00000005, 0x00000005}},
+            {0x00001f43, new uint[] {0x00000005, 0x00000005, 0x00000005}},
+            {0x00001f44, new uint[] {0x00000005, 0x00000005, 0x00000005}},
+            {0x00000fa6, new uint[] {0x00000000, 0x00000001, 0x00000007}},
+            {0x000007d7, new uint[] {0x0000000A, 0x00000008, 0x00000007}},
+            {0x000061a8, new uint[] {0x0000000B, 0x0000000C, 0x0000000C}},
+            {0x000061a9, new uint[] {0x00000007, 0x00000004, 0x00000002}},
+            {0x000061aa, new uint[] {0x0000000C, 0x00000007, 0x00000009}},
+            {0x00006978, new uint[] {0x00000002, 0x00000004, 0x00000009}},
+            {0x0000697c, new uint[] {0x00000005, 0x00000003, 0x00000006}},
+            {0x00006d60, new uint[] {0x0000000C, 0x00000009, 0x00000007}}
+        };
+
+        private TCPGecko Gecko { get; set; }
+        private uint Offset { get; set; }
+
+        public GearUnlocker(TCPGecko gecko, uint offset)
+        {
+            Gecko = gecko;
+            Offset = offset;
+        }
+
+        public void PokeClothes()
+        {
+            PokeGear((uint)GearAddress.Clothes + Offset, _clothes);
+        }
+        public void PokeHats()
+        {
+            PokeGear((uint)GearAddress.Hats + Offset, _hats);
+        }
+
+        public void PokeShoes()
+        {
+            PokeGear((uint)GearAddress.Shoes + Offset, _shoes);
+        }
+
+        private void PokeGear(uint baseAddress, Dictionary<uint, uint[]> gear)
+        {
+            // Sort the Dictionary's keys so that starter gear will appear first
+            var sortedKeys = gear.Keys.ToList();
+            sortedKeys.Sort();
+
+            foreach (var objectId in sortedKeys)
+            {
+                var abilities = gear[objectId];
+
+                // Poke the memory addresses
+                Gecko.poke(baseAddress, objectId); // objectId
+                Gecko.poke(baseAddress + 0x00000004, 0x00000004); // level
+                Gecko.poke(baseAddress + 0x00000008, 0x00000004); // slots
+                Gecko.poke(baseAddress + 0x0000000C, abilities[0]); // slot 1
+                Gecko.poke(baseAddress + 0x00000010, abilities[1]); // slot 2
+                Gecko.poke(baseAddress + 0x00000014, abilities[2]); // slot 3
+                Gecko.poke(baseAddress + 0x00000024, 0x00000024); // date
+                Gecko.poke(baseAddress + 0x00000028, 0x00010000); // new flag
+
+                // Move to next gear slot in the inventory
+                baseAddress += 0x00000030;
+
+                // debug
+                // Console.WriteLine("poked (objectId = " + objectId + ", new baseAddress = " + baseAddress + ")");
+            }
+
+            /*if (SendStats) ToDo 
+                StatisticTransmitter.WriteToSlot(10, 1);*/
+        }
+    }
+}
