+++ SplatAIO/Logic/Hacks/Unlock/MinigamesUnlocker.cs
@@ -1,0 +1,6 @@
+﻿namespace SplatAIO.Logic.Hacks.Unlock
+{
+    public class MinigamesUnlocker
+    {
+    }
+}
