--- SplatAIO/Logic/Memory/Addresses/Sisterhax.cs
@@ -1,8 +1,0 @@
-﻿namespace SplatAIO.Logic.Memory.Addresses
-{
-    public enum Sisterhax : uint
-    {
-        Aori = 0x105EB5EC,
-        Hotaru = Aori + 0xC //0x105EB5F8
-    }
-}
